+++ mj-calc/src/modules/hooks/useLocalStorage.ts
